requirements.txt
yt-dlp
google-genai
python-dotenv~=1.1.0
platformdirs~=4.3.8
